Try to parse into a known error on fail.
Get the body text and try to parse into a login response.
If that fails try to parse the body to get an `error_description` from
Azure. If that fails report whatever the original issue was with parsing
as the error.

`ErrorResponse` could be expanded on to recognize the `error` field and
give more useful info. Not sure what connecting `ErrorResponse` to
`AzureError` in a nice way looks like, but this leaves all the
interfaces the same for now.