Use endpoint URIs from connection string in storage KeyClient
Endpoint URIs can be specified in the storage connection string.
Currently they are ignored and replaced with hard-coded default values
pointing to the azure blob storage.

This is relevant when using the azurite storage emulator, where the account
names can be modified (since version 3). The "with_emulator" method with hard
coded value is no longer sufficient if account names are customized.