Expose azurite_workaround feature in storage crate and use endpoint URIs from connection string in storage KeyClient (#101)
* Expose azurite_workaround feature in storage crate

The azurite storage emulator requires specific configurations that are
already available in the source code of the storage crate.
The feature is not exposed, to be used though.

* Use endpoint URIs from connection string in storage KeyClient

Endpoint URIs can be specified in the storage connection string.
Currently they are ignored and replaced with hard-coded default values
pointing to the azure blob storage.

This is relevant when using the azurite storage emulator, where the account
names can be modified (since version 3). The "with_emulator" method with hard
coded value is no longer sufficient if account names are customized.