Compile sdk/core and sdk/storage to wasm32-wasi
This commit adds a new HTTP client in sdk/core under a feature that
allows compilation to wasm32-wasi, using the
wasi_experimental_http library.

While this is a proof of concept, and other SDKs could be updated in this
way, not all of them will be compatible - specifically, those that
require Hyper or Reqwest without going through sdk/core.
They could also be updated, but depending on the type of functionality
required, some might not work.

Signed-off-by: Radu M <root@radu.sh>