🐛: Fixes metadata for list blobs not populated
This is a tricky one, the XML response from Azure for a list of blobs
holds the metadata next to the properties, but the Blob struct is
build to hold the metadata inside the BlobProperties. This moves
the metadata field from the BlobProperties to the Blob struct, but
this would break a public interface of the crate.

An example XML response, when metadata is included:

```
<?xml version=\"1.0\" encoding=\"utf-8\"?>
<EnumerationResults ServiceEndpoint=\"https://....blob.core.windows.net/\" ContainerName=\"...\">
    <Prefix>../</Prefix>
    <Delimiter>/</Delimiter>
    <Blobs>
        <Blob>
            <Name>..</Name>
            <Properties>
                <Creation-Time>...</Creation-Time>
                ....
            </Properties>
            <Metadata>
                <someKey>someValue</someKey>
            </Metadata>
            <OrMetadata />
        </Blob>
    </Blobs>
    <NextMarker />
</EnumerationResults>
```

A better approach would be to tell serde to deserialize it correctly, but I'm not
sure how to do this.

Change-type: patch
Signed-off-by: Andreas Fitzek <andreas.fitzek@reactivereality.com>