set content-type application/json for generated requests with a json body
Muliple Azure services reject requests without an appropriately set
content-type (Application Insights) is one such.

Long term, this should check the `consumes` part of the endpoint spec,
which every spec I spot checked, if it has it defined, is
`["application/json"]`.

This PR sets the content-type based on the assumtpion that "any time the
BODY is json encoded, we should set the content-type"