work around applicationinsights data plane spec error
The Application Insights dataplane spec states that table.row should be Vec<Vec<String>>.  However, in practice can be other data types.  I've personally experienced JSON `int` as well as `string`.

This force sets the data type to a serde_json::Value.  While this is less optimal, it means the SDK is functional while waiting on the spec to be fixed.

This PR includes a working example that executes a query and naively renders the results.

Ref: https://github.com/Azure/azure-rest-api-specs/issues/16425