pass Headers as part of prepare_request
prepare_request adds the authentication related headers.  When using
Access Key methods, both Blob Storage and Data Tables uses data from the
headers as part of canonicalizized data that gets signed, which is
checked at the server sidej.

As such, the headers must be part of what is sent to prepare_request.

See #871 for a walk through for Data Tables.