blob_storage: Make `Snapshot` type consistent.
Previously many things treated `Snapshot` as an opaque type defined by
the `request_query` macro, but the `Blob` type was deserializing it as a
`DateTime` and not reading it from the header at all. Now everything is
using one shared opaque type.

I chose to redefine `Snapshot` without using the macro so that it could
derive Serialize/Deserialize. This is required because in the cases
`Snapshot` is part of a serialized response body. It is otherwise compatible with the
previous definition. The alternative would have been to change the macro
to do derive Serialize/Deserialize, but since that would be a change to
`core` and a number of other types, it felt heavy handed.