reduce log level on requests that will not be retried
This log entry generates errors on multiple common patterns that should not result in the SDK generating error logs.

As an example, the recommended method for checking if a blob exists is to read it's properties.  If a blob does not exist, it generates a 404 response, which causes an error here.

Given the SDK returns an error, this should be left up to the SDK consumer if this warrants logging an error.