actively avoid panic at runtime in keyvault crate
This PR updates a handful of operations to not use `unwrap` in the
operations and adds clippy configuration to deny the use of `unwrap` or
`expect` (with the exception of unit tests) in the crate moving forward.

This also adds a top level clippy.toml that explicitly allows `unwrap`
and `expect` when used within `#[cfg(test)]` blocks.