remove rust-toolchain.toml (#1320)
Instead of specifying the toolchain via `rust-toolchain.toml`, our CI pipelines specify the intended toolchain.  For contributor's purposes, this is out of sync with what we verify against for PRs and has contributed to confusion over time.

For users of the SDK, this isn't applied.  For those, we specify the toolchain needed via setting MSRV in the `Cargo.toml` for the specific crate.

Co-authored-by: Brian Caswell <bcaswell@microsoft.com>