unify management of MSRV
This PR standardizes where all crates set the MSRV (azure_core), as well as updates the actions to build on a matrix of 'stable', 'nightly', and the rust-version set in `azure_core`.

This allows developers of the SDK to run each of the build tasks exactly as they are run in CICD with minimal configuration.