at the time of Pipeline::send, context doesn't need to be mutable (#1355)
* at the time of Pipeline::send, context doesn't need to be mutable

Up until we get to `Pipeline::send`, Context can (and likely will) be
mutable.  However, once we start processing the pipeline, it doesn't
need to be mutable anymore.  This is due to `Policy::send` not requiring
mutable context.

The changes to `azure_iot_hub` and `azure_security_keyvault` are related
to longer requiring mutable contexts once we get to `Policy::send`.

Ref: https://github.com/Azure/azure-sdk-for-rust/blob/43d4e9974060e78bbd8912ddcc183567781997fc/sdk/core/src/policies/mod.rs#L30-L33