begin LRO implementation for generated crates
This is an initial implementation for supporting `properties.provisioningState` model of long running operations.  This models the Python SDK for how the state is extracted.

Once each of the methods that identifying state are included, thish should be wrapped up into a trait that the clients can implement, rather than manually implementing the polling in the generated `into_future`.