add `models` to autorust's reserved words
Because we use `models` heavily throughout to describe where the
generated structs are stored, services that use `models` as a identifier
name will cause issues.

By adding `models` to the pre-existing `is_keyword` and using a
something that doesn't check `is_keyword` for when we _actually_ want
`models`, we can generate code for services that use the keyword
`models`.

Fixes #1382