fix timezone conversion
The original change in #1431 was built & tested on a machine with the
timezone of GMT, which is effectively UTC as it relates to the testing.

In addition to fixing the conversion, this adds a unit test for unix
platforms that temporarily changes the `TZ` environment variable (used
for setting the timezone for libc platforms), performs the conversion,
then resets the `TZ` environment variable.

This validates the functionality regardless of the testing environment.