expose the default tag as a feature
Right now, if you want to specify the http client implementation, you
_also_ must specify the package tag which can change regularly during
updates.

This moves to including a feature named `default_tag` which refers to
the specific tag name, as well as renames `no-default-tag` to
`without_tag_import` to do what was previously referred as
`no-default-tag`.

This allows users to have a consistent Cargo.toml entry for the latest
tag and a specified http client such as the following:

`azure_svc_batch = {version="0.17", default-features=false, features=["enable_reqwest", "default_tag"]}`