standardize on xml & json helpers
In past PRs, we added centralized json & xml helpers.

This builds upon that idea by exposing helpers for Request & Response
(including CollectedResponse and ResponseBody, as those are sometimes
needed as well).

This significantly reduces the complexity for _most_ of the callers and
models what `reqwest` does for handling serialization and
deserialization of requests.