Make consistent with other Azure SDK languages
This also turns the `source_id` (nee `id`) and `vault_url` (new) into owned `String`s because we don't really intend customers to do anything with them as `Url`s. If they really need a `Url` from it they can parse it into one, but there are really no supported scenarios for this for callers of this crate. The `endpoint` parameter - the `vault_url` here - `name`, and `version` parameters are all `&str` slices.