Add resource_id() extension method to Secrets (#2052)
* Add resource_id() extension method to Secrets

Parsing the name anv version from the resource ID (URL) is something we do across other Azure SDK languages and makes the API much more user-friendly. It's common to need to parse the name and version from a secret (or key or certificate) and pass into subsequent calls.

* Make consistent with other Azure SDK languages

This also turns the `source_id` (nee `id`) and `vault_url` (new) into owned `String`s because we don't really intend customers to do anything with them as `Url`s. If they really need a `Url` from it they can parse it into one, but there are really no supported scenarios for this for callers of this crate. The `endpoint` parameter - the `vault_url` here - `name`, and `version` parameters are all `&str` slices.