Automatically record and play back test recordings (#2033)
* Automatically record and play back test recordings

Resolves #1876

* Require `Result<T, E>` from recorded tests

This is simpler to support in the `#[recorded::test]` macro and sets a good precedent we want to show users anyway instead of `unwrap()` or `expect()`.

* Fix azure_core_test_macro tests

* Switch to multi-threaded async runtime and send correct headers

The default single-threaded async runtime runner was deadlocking in `Recording::drop()`. I opened #2049 to revisit this but I don't foresee any problem. We already don't support test-proxy on wasm32 and our mix of agents we do test on support multi-threaded async runtimes just fine.

* Fix Windows tests

* Safely log headers and body

We need to sanitize the headers later, but easier just to sanitize them all right now. This is still better than before which only showed how many headers were in the request.

* Fix regression to auto-start of test-proxy

* Resolve PR feedback