Use specific versions of litemap, zerofrom
Only in the `rust - pullrequest` pipelines was it seemingly ignoring our `Cargo.lock` file and using newer versions of these crates that upgraded their MSRV.
This explicitly adds the versions we had in our `Cargo.lock` file to see if that passes the PR builds. We may still want to pass `--locked` in both PR and CI builds;
however, oddly, the CI builds are passing even after a manual run I just kicked off an hour or so ago.