Bump versions and simplify dependencies (#2294)
* Bump versions and simplify dependencies

At least in the workspace `Cargo.toml`, crates that should only ever be in `dev-dependencies` can have a `path` and no `version`. This makes referencing them easier, but I'd be fine with crates taking a local `path` as well.

* Remove azure_identity dependency from storage

* Use specific versions of litemap, zerofrom

Only in the `rust - pullrequest` pipelines was it seemingly ignoring our `Cargo.lock` file and using newer versions of these crates that upgraded their MSRV.
This explicitly adds the versions we had in our `Cargo.lock` file to see if that passes the PR builds. We may still want to pass `--locked` in both PR and CI builds;
however, oddly, the CI builds are passing even after a manual run I just kicked off an hour or so ago.

* Update crate names, used --locked

Resolves #2300

* Fix typespec_macro compilation tests

The Cargo.lock file wasn't committed; still, don't know why it just started failing in PR but still not in CI builds.

* Remove `--locked` from build and test commands

We'll decide as part of #2300. Kept it in `package` and `publish` to make sure we ship what we intended.