Remove parsing module
The parsing module only contained the `FromStringOptional` trait implemented for a few primitives and implemented by our `create_enum` macros. But it's not used, and I could find no use of it even in the `legacy` branch. Maybe it once was, but it really is just `FromStr` now with `FromStr::Err` set to `typespec::error::Error`. No need to keep it, and it just creates problems for re-exporting macros into `azure_core` without necessitating a dependency on `typespec_client_core`.