Implement Recoverable connection for AMQP operations. (#2657)
This PR does three different things:

It renames the ConnectionManager type to RecoverableConnection to better reflect what is going on.
It splits out the authorization functionality from ConnectionManager into a new struct named Authorizer to simplify the logic in connection_manager.rs
It consolidates all the AMQP active protocol interactions to RecoverableConnection. This was done to (a) simplify the AMQP interactions for each of the specific protocol elements and (b) lay the foundation for connection, session, and link recovery. See the rather large comment on the RecoverableConnection for more details on how this all works and fits together.
There are a few bonus changes in this PR as well:

Adds a send_ref function to the AmqpSender struct which allows sending a reference to a message (saving a message clone).
AmqpOrderedMap::iter() passes references to the inner elements to be consistent with the standard collection iter() method.