WIP: Support custom serialization for RequestContent
This doesn't work for `XmlFormat`, though. A generic `SerializeWith<XmlFormat>` implementation for some arbitrary type - container or element - can't really know if it's a document or fragment.

But, the main goal here was to reduce the boilerplate an emitter - or dev writing a convenience layer - has to write. It didn't reduce it much anyway.

I'm going to open a separate PR with just the changes to `RequestContent` to add the `F: Format` type parameter so that we have a client-defined format for both serialization and deserialization.