Adding Location Cache (#2801)
### PR Summary:
* Location cache is responsible for organizing the database account data
gathered by the global endpoint manager. It holds the client’s preferred
locations as well as lists of write and read regions separated by
location and availability. Its purpose is to provide an available
endpoint based on the client's preferred locations.
* Added location cache functionality and tests. 

### Implementation details:
* APIs include: 
  - update
  - refresh_endpoints
  - refresh_stale_endpoints
  - get_endpoints_by_location
  - get_preferred_available_endpoints
  - mark_endpoint_unavailable
  - is_endpoint_unavailable
  - resolve_service_endpoint

#### Current limitations:
* Currently does not support multi-write capabilities, planning to add
it in a follow-up pr with global endpoint manager.

---------

Co-authored-by: Ashley Stanton-Nurse <ashleyst@microsoft.com>