Added ability to test distributed tracing functionality. (#2896)
Fixes #2769 - Adds ability to easily test distributed tracing.

Also adds a test example for KeyVault Secrets APIs.

This PR introduces a new `azure_core_test` API named
`assert_instrumentation_information` whose purpose is to ensure that the
distributed tracing information is correctly generated for a given
service client implementation.

The API takes three pieces of information: A capture which creates a
service client with a provided distributed tracing `TracerProvider`, a
capture which actually performs the test and an
`InstrumentationInformation` structure which describes the service
calling the API and the particular service client APIs which are
expected to have been called during the test.

It then analyzes the actual traces generated during the test and
verifies that the expected service client calls were made and that the
expected distributed tracing attributes have been generated.

---------

Co-authored-by: Heath Stewart <heaths@microsoft.com>