Add a Logging policy with sanitized HTTP headers and URLs (#2955)
Also updated the `Debug` trait implementation for `Request`,
`RawResponse`, and several other locations that print URL types to use
the default sanitization rules.

Sanitization is done using an "allow list" - the customer (and/or a
service) can provide an additional set of headers and/or URL query
parameters to be allowed, but by default most query parameters and
headers are redacted when logged.

As a bonus, this trims the public API surface associated with HTTP
headers to the set of headers which is currently in use by the SDK
internals and other service clients.