cosmos: Changelog prep for 0.28 release (#3116)
This PR preps the CHANGELOG.md for the 0.28 release. There are three
separate changes here, which I put in separate commits, because the
overall diff is a little busier than you might expect :).

1. Added a missing changelog entry for #2969. I forgot to ensure that
was added before merging, so I put it in here.
2. Replaced our usage of `*` in the changelog with `-` so we're
consistent with other SDKs, and consistent within the file (since a
recent Core change added a changelog entry using `-`)
3. Marked the changelog entry for `0.28.0` as releasing tomorrow.