cosmos: add experimental c wrapper around cosmos sdk (#2906)
We want to explore the option of publishing a C library of the Azure
Cosmos DB SDK for Rust. This PR adds the basic framework to start that
effort:

* A new package `cosmosclient` that depends on `azure_data_cosmos`. This
package builds a `cdylib` and `staticlib` target
(`libcosmosclient.a`/`libcosmosclient.so`/`cosmosclient.dll`/etc.) that
exports a C-compatible API. For now, that API is extremely simple
(`cosmosclient_version`).
* A simple test infrastructure (NOT currently integrated into anything).
There are C programs in `sdk/cosmos/cosmosclient/c_tests` that will be
built and linked against the cosmosclient library. The tests are simple
pass/fail programs. If the test returns a `0` exit code, it passed,
otherwise, it failed.
* `build.rs` script in `cosmosclient` to produce
`sdk/cosmos/cosmosclient/include/cosmosclient.h`, which is a C header
file defining the API surface of `cosmosclient`. We should publish this
as an artifact at some point, but for now it's there to be used by
tests.