Adds new AAD scope value instead of account scope (#3246)
As part of this PR we are adding a new AAD audience scope value
"https://cosmos.azure.com/.default". We will be using this instead of
account scope value in rust sdk. This new scope value will be eventually
used across different clouds and regions. It is already accepted in
public cloud across stage, canary and prod.