Use custom error for AMQP and EventHubs (#3248)
Define a new custom error type for the `azure_core_amqp` (`AmqpError`)
and `azure_messaging_eventhubs` (`EventHubsError`) packages.

Ensured that `azure_core::error::Error` types can be embedded in AMQP
and Event Hubs errors, and that a `try_from` implementation exists to
convert back to Azure Core errors (the `try_from` fails if the
AMQP/EventHub error is not derived from an Azure Core error).

Also added `AmqpError::with_message` and `EventHubsError::with_message`
to cover the cases where a simple error message needs to be returned to
the caller.

Also refactored the AMQP "error" module into 3 source files:`mod.rs`
which implements the module, `tests.rs` which implements test cases, and
`error_condition.rs` which implements the `AmqpErrorCondition`
extensible enumeration - this was split out to reduce the size of the
`mod.rs` file.


Note: This turned out to be somewhat more complicated than I thought it
would be originally.

The core problem is that instead of having a single error type for
everything, we have 3 different error types, each of which is used in
different circumstances. That made figuring out which error kind to
return from which function rather challenging.

Specifically in the EventHubs code which implements AMQP functionality
(including the logic which creates AMQP structures), the functions need
to return an *AMQP* error code even though they are in the EventHubs
package, and are interacting with EventHubs structures - that's because
ultimately the code which consumes these structures assumes that it will
receive an AMQP error, NOT an EventHubs error.

Bonus: Cleaned up a bunch of unused EventHubs errors.