cosmos: Allow the Query Pipeline to return an alternative query to execute in each request (#3166)
For Hybrid Search, as well as ReadMany support, the query engine will
need to direct the SDK to issue **multiple different queries**. However,
the Rust SDK's query executor expects to be able to replay the original
query on each request. This PR changes that expectation to allow (but
not require) each Query Request coming from the pipeline to provide an
alternate query. When an alternate query is provided, the pipeline can
also indicate if the parameters from the original query should be sent.
Often, the alternate query does NOT use the parameters, but in some
circumstances (hybrid search) it will.

There are also some small fixes for doc tests and error checking in the
engine-less cross-partition query tests. Not sure why these are coming
up here and not in `main`.