cosmos: remove unnecessary dependency on typespec_client_core (#3260)
We had a direct dependency from `azure_data_cosmos` to
`typespec_client_core`. That was only necessary back when we had the
`#[derive(Model)]` macro, because it required that direct dependency.
Now that we've removed that macro, we don't need the
`typespec_client_core` dependency. All the types we use are re-exported
from `azure_core`, and in fact we use them through those re-exports. So,
we can remove this dependency now.