cosmos: updates to the query engine interface for hybrid search (#3275)
Two small changes to the query engine interface based on work I'm doing
to prep for Hybrid Search support:

1. Allow the pipeline to return a `None` query, indicating that
**every** request will come with an override query. We validate that
with a panic in dev builds, but return an `Err` in release builds.

2. Update the `index` field on `QueryRequest` to be called `id`, to more
accurately describe that it's an opaque ID rather than a sequential
index. In Hybrid search, we encode things in this ID like the specific
component query the data represents and that's not really an ordered
concept as far as the SDK is concerned. This number is just basically
just a token that the engine provides to the SDK which must be copied
across to the response so we can correlate them.