Implement Pager resumption even through into_pages (#3286)
By design, we didn't support resuming `Pager` (`ItemIterator`) but, upon
further discussion, decided we can but will resume from the current page
until all items on the current page are processed.
    
This also fixes a problem that, because of the design, we didn't pass
through resumption to the `PageIterator` at all - it couldn't even
resume as a `PageIterator` for subsequent pages.