Moved otel support for polling into poller (#3279)
Removed custom code from `create_certificate` API code and moved the
corresponding logic into the `Poller`.

This involved adding an additional parameter to the `callback` for
`Poller::with_callback` that provides the `ctx` used for the pipeline.

Bonus: Instead of using a tuple to maintain the state of the "stream"
created by the `Poller`, encapsulate the state in a structure to make it
easier to modify the state of the "stream" moving forward (when there
were 3 fields in the state, it wasn't a huge deal, but with 5 fields, it
was getting unwieldy).

NOTE: This is still a draft PR, the next step is to work on similar
changes for pagers.