Cosmos: Adds `GlobalEndpointManager`, `ClientRetryPolicy` and `LocationCache` To Support Cross-Regional Routing (#3348)
## Description:
 
This PR is introducing `GlobalEndpointManager`, `ClientRetryPolicy`,
`LocationCache` and `AccountProperties` to support cross-regional
routing in the rust SDK. Some of the key design decisions are
highlighted below:

- We used [Moka ](https://docs.rs/moka/latest/moka/)as a caching library
which supports thread safe access of the cached data and uses LRFU based
cache eviction mechanism.

- The cache refresh has been done on a TTL (time to live) model, where
each incoming request has the opportunity to refresh `AccountProperties`
and `LocationCache`, if the TTL expires or `force_refresh` is passed as
`True`.

- We haven't took the background refresh route yet, since we will need
to be a little careful and optimized for the rust bg task management.
However, our long term goal is to refresh the cache in the background,
so that we do not penalize an incoming request to refresh the cache.

The below flow diagram depicts the entire component interactions to
better understand how the request will be made.

## Flow Diagram:

```mermaid
sequenceDiagram
    participant J as ContainerClient
    participant A as CosmosPipeline
    participant B as RetryHandler
    participant C as ClientRetryPolicy
    participant D as Core Pipeline <br> send_raw <br> through delegate
    participant E as GlobalEndpointManager
    participant F as LocationCache


    J-->>A: 1. send(CosmosRequest)
    A->>B: 2. send<Sender, Fut>()
    loop Retry Iterations with force-refresh flag = false/ true
        B->>C: 3. before_send_request(request)
        C->>E: 4. refresh_location_async() <br> Refresh Cache only when TTL Expires OR <br> force_refresh is requested. <br><br> NO-OP when either of the above is false.
        E->>E: 5. get_database_account() <br> Fetch Account Properties <br> from Gateway.
        E->>F: 6. on_database_account_read() <br> Refresh Location Cache
        C->>E: 7. resolve_service_endpoint() <br> Fetch and set the next endpoint in <br> CosmosRequest to route request.
        B->>D: 8. Call sender delegate(request)
        B->>C: 9. Invoke should_retry(result)
        B->>B: 10. RetryAfter <br> (TimeSpan.Zero)
    end
        B->>J: 11. Return result if No Retry
```

## Closing issues

Fixes Azure/azure-sdk-for-rust#3175