cosmos: Fix URL encoding for authentication signatures (#3461)
This PR fixes two bugs, with the same root cause: We URL encode paths
before signing the request, but we should **not** be doing that.

* We have occasional test failures related to throughput offers when the
offer ID has a RID that includes `+`, because we encode it
* #3394 - Whenever an ID has a URL-unsafe character in it, we produce an
invalid resource link for signing and the user gets a 401

This PR fixes both by storing both the unencoded and encoded form of
path segments in the `ResourceLink`, which get preserved as links are
combined (to create sub-clients). In some SDKs, we "unencode" the string
before generating the signature payload, but that's challenging because
the current URL encoding library we're using (`form_urlencoded`) doesn't
actually support decoding. We can improve this later, for now I just
want to get something that fixes the immediate issue.

We don't have direct APIs for managing offers, so the easiest way to
end-to-end test this was to produce a test that kept trying to create a
database until it got one with an offer RID containing `+`. I did that,
saw it fail, fixed this issue, and it passed. Since that test is
inherently inconsistent, I haven't actually included it.

Closes #3394